fy/code/prod/requirements.txt
# pip3 install mysql-connector-python
mysql-connector-python==8.0.33
numpy==1.21.4
redis==4.5.4